Significant Refactoring: Signals like SegmentEndPointSignal used to carry Data inside it. Now, the Data has been separated from the Signal, so a Signal is purely a Signal. Consequently, the Data/Signal stream would now looks like:
<SegmentEndPointSignal> <AudioFrame> ... <AudioFrame> <SegmentEndPointSignal>

instead of

<SegmentEndPointSignal> <AudioFrame>...<AudioFrame> <SegmentEndPointSignal>
   |-> <AudioFrame>                                    |-> <AudioFrame>

Not only is this approach conceptually better, but also (contrary to my
initial believe) makes the implementation of the processors a lot cleaner.
There is a lot less type casting and requires no indirection
to get to actual Data object inside the Signal. I'm happy with this
refactoring :) Although it requires some look ahead in the FeatureExtractor
(which is handled by a peekable InputQueue), I believe this is the right
approached.

All previous tests passed.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@323 94700074-3cef-4d97-a70e-9c8c206c02f5