(1) Added Live mode AudioSource (just as a trial). It uses the Java Sound API to capture audio from the system's default audio capturing device. It buffers all the audio data in between startRecording() and stopRecording(), as we discussed in the email (yes, its an infinite cache, but we can change that).
(2) Added a swing test program for the live-mode AudioSource. You can
try running it at the tests/frontend directory by typing "gmake live".
The code is really preliminary and needs a lot of work. A swing app. will
show up, with a "Listen" and "stop" button. You need to have a microphone
attached to your computer. The features produced from the captured audio
will be shown. Please report any problems/crashes,
I'm using it to debug the frontend in live-mode.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@424 94700074-3cef-4d97-a70e-9c8c206c02f5