Added file that collects interesting files from the regression test directory and archives them in compressed form, so that debugging of failed tests becomes easier.
Archives overwrite previous archives with the same name, and are
stored as /lab/speech/sphinx4/regressionArchive/TEST.MACHINE.zip,
where TEST and MACHINE depend on, well, test and machine. This way,
nightly tests are overwritten everyday, and other tests are
overwritten once a week.

The files that are stored in the compressed files are files named
"*.out", if their modification date is less than a week old,
"hs*.log", and "gc.txt". Files are "zipped".


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@2037 94700074-3cef-4d97-a70e-9c8c206c02f5