Modified the makefile to allow builds that depend on the jsapi.jar to proceed by attempting to copy the jsapi.jar from an external location. This allows us to add code to the source tree that depends on jsapi.jar without breaking the nightly builds on known systems.
Nightly build systems (glottis, mickey, mangueira) should have the
jsapi.jar placed in /lab/speech/sphinx4/lib/jsap.jar.  (I've done this
already).

(Note, we are switching over to ant soon, so the equivalent target
should be added to the build.xml file)

The details:

Augmented the 'make all' target to install the jsapi.jar into the lib
directory from an external location if necessary.
If the file sphinx4/lib/jsapi.jar already exists, nothing is done,
however if it does not exist, an attempt is made to copy a jsapi.jar
from the external directory /lab/speech/sphinx4/lib to ./lib.  If the
external directory does not exist then an error is generated with
instructions for the user to unpack the jsapi.jar file in the lib
directory.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@2338 94700074-3cef-4d97-a70e-9c8c206c02f5