The language model keeps a cache of trigram and bigram probabilities. This cache was only cleared between utterances.  For some particulary long utterances or when certain LM experiments were performed this unbounded caching was, in effect, a memory leak.
I've modified the bigram and trigram caches to be bounded. The bounds
are determined by the sphinx properties:

    PROP_TRIGRAM_CACHE_SIZE  and PROP_BIGRAM_CACHE_SIZE

These caches use an LRU replacement policy.

The default sizes for the cache are such that behavior for utterances
of 30 seconds or less is identical to the unbounded cache behavior.

Since the caches are now bounded, it is no longer necessary to clear
the caches between utterances. This may improve performance in the case
where consecutive utterances have similar word sequences.  This is now
the default behavior.

The sphinx property:

    PROP_CLEAR_CACHES_AFTER_UTTERANCE

controls this behavior.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@2452 94700074-3cef-4d97-a70e-9c8c206c02f5