Merged from branch SPHINX3_5_2_RCI_IRII_BRANCH:
1, Added Letter-to-sound LTS rule, dict_init will only specify
d->lts_rules to be true if the useLTS is specified.  Only if
d->lts_rules is specified, the LTS logic will be used. The code safe
guarded the case when a phone in mdef doesn't appear in LTS, in that
case, the code will force exit.

2, The LTS logic is only used as a reserved measure.  By default, it
is not turned on.  See also the comment in kbcore.c and the default
parameters in revision 1.3 cmdln_macro.h . We added it because we have
this functionality in SphinxTrain.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@5283 94700074-3cef-4d97-a70e-9c8c206c02f5