Sync'ing wiht SphinxTrain.
Added the flag "-seed". If dither is being used and the seed is less
than zero, the random number generator is initialized with time(). If
it is at least zero, it's initialized with the provided seed. This way
we have the benefit of having dither, and the benefit of being
repeatable.

This is consistent with what sphinx3 does. Well, almost. The random
number generator is still what the compiler provides.

Also, moved fe_init_params to fe_interface.c, so one can initialize a
variable of type param_t with meaningful values.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@5396 94700074-3cef-4d97-a70e-9c8c206c02f5