Bring the allphone decoder in from the cold, so to speak.  It now uses fast GMM computation and more of the standard stuff, and this seems to have made it both faster and more accurate.
It is not yet implemented as a "search mode".  However, this should
not be hard to do in the near future.  At the very least, the
functions in s3_allphone.c are going to get integrated into the main
Sphinx3 library so that the Python interface (which should be
committed at some point soon) can get to them.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@6651 94700074-3cef-4d97-a70e-9c8c206c02f5