Resurrect FAST8B, albeit in a more general and (hopefully) more obvious and documented form.  The key to making this fast is being able to avoid all of the safety checks that the real logmath_add() does.  We achieve this by ensuring that the acoustic scores are normalized (so they are always negative) and that the magnitude of the sum of the log acoustic scores and log mixture weights never exceeds 255.  Therefore we can use a fixed size logadd table.
The old code used stupid magic numbers like -99000 and -162000 here
and probably had off-by-one errors.  It also made the log-add table
WAY bigger than it needed to be by treating it as a 2D array rather
than using the difference between the two input numbers.  This is
theoretically a bit slower but it allows us to use a table of 30 bytes
rather than 64 kilobytes.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@7370 94700074-3cef-4d97-a70e-9c8c206c02f5