Decouple lattice search from N-Gram search.
Though, it's worth noting that the lattice search actually does use
N-Gram models, so it's not totally clear what this gains for us :)

Oh, actually, it means this stuff can go in SphinxBase.  Which it
will, probably kind of soon, along with a lattice to FSG converter
which we would like to have for discriminative training.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@8097 94700074-3cef-4d97-a70e-9c8c206c02f5