lm3g_tginfo_reset() wasn't actually used, leading to memory "leaks" (though all memory was accounted for) due to the tginfo cache never being flushed.
This patch makes it accessible from the external API, and also makes
it more agressive (and probably faster) by just freeing the whole
allocator.

A better strategy particularly for WinCE or for MMU-less systems would
be to have a fixed size trigram cache and some strategy for expiring
entries once it fills up.


git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@8793 94700074-3cef-4d97-a70e-9c8c206c02f5