Added an example Groovy script.  The script is located in a new directory src/scripts/location
Here is the README

This directory contains a example Groovy script.  This example reimplements the Transciber application. 

Groovy is a popular choice for Java based scripting, and is a good choice for rapid development with Sphinx4.

To try this script, install Groovy from here

    http://groovy.codehaus.org/

Then cd to src/scripts/groovy and type the following to the shell

    groovy -cp ../../../lib/sphinx4.jar:../../../lib/jsapi.jar GroovyTranscriber.groovy ../../../




git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@9416 94700074-3cef-4d97-a70e-9c8c206c02f5